requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
Pillow>=9.0.0
uuid
speechrecognition>=3.10.0
pyaudio>=0.2.11
gtts
pydub